I made these changes. I resized the main window so every information displays correctly. In the main window, i added some css to the menu bar so that is easier to see. In the FreePlayWindow i added some a pane on the top of the main pane which contains a text that pops up when the user presses one note - and shows the pressed note. This pane also contains two new buttons. One ToggleButton which shows notes on each key when it is pressed and a HomeButton that is used to return to the main window(Better Navigation).
In LearnWindow i also added home button (returns to main window) and changed a little bit the display of keypane. I put keypane on the center of an Hbox which also contains a text that shows the sequence of notes weather it is about Chords or Scales on the Left.