UI improvements (#46)
* I made these changes.
I resized the main window so every information displays correctly.
In the main window, i added some css to the menu bar so that is easier to see.
In the FreePlayWindow i added some a pane on the top of the main pane which contains a text that pops up when the user presses one note - and shows the pressed note. This pane also contains two new buttons. One ToggleButton which shows notes on each key when it is pressed and a HomeButton that is used to return to the main window(Better Navigation).

In LearnWindow i also added home button (returns to main window) and changed a little bit the display of keypane. I put keypane on the center of an Hbox which also contains a text that shows the sequence of notes weather it is about Chords or Scales on the Left.

* In this commit, I made the changes mentioned on the FreePlayWindow. Such changes are: a) the removal of label that shows note pressed -  b) The text that appears when the user pressed ToggleButton showNotes. I reduced the size of text in black keys so that it is on one line and also get rid of the numbers that shows in which octave the button is since it is not practical.

* All gui centered in mainWindow with added css on top pane that contains menu bar

* All gui centered in mainWindow with added css on top pane that contains menu bar

* Delete sounds directory

Signed-off-by: NikosVogiatzis <92678797+NikosVogiatzis@users.noreply.github.com>

* Removal of text showing sequence of buttons pressed on a chord or scale action. Added: the display of note names on keyboard only of those notes that are contained on scale/chord.

---------

Signed-off-by: NikosVogiatzis <92678797+NikosVogiatzis@users.noreply.github.com>
Co-authored-by: Nikolaos Vogiatzis <vogiatzisnik02@gmail.com>